# docs/openapi/v1.0-api-contract.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/buildings/setup/draft": {
      "post": {
        "operationId": "BuildingController_saveDraft_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      },
      "get": {
        "operationId": "BuildingController_resumeDraft_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/setup/submit": {
      "post": {
        "operationId": "BuildingController_submit_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/lookup": {
      "get": {
        "operationId": "BuildingController_lookupPostalCode_v1",
        "parameters": [
          {
            "name": "postalCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings": {
      "get": {
        "operationId": "BuildingController_listMine_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}": {
      "get": {
        "operationId": "BuildingController_getOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/units": {
      "get": {
        "operationId": "BuildingController_listUnits_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      },
      "post": {
        "operationId": "BuildingController_addUnit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/unit-metadata": {
      "get": {
        "operationId": "BuildingController_listUnitMetadata_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/units/{unitId}": {
      "get": {
        "operationId": "BuildingController_getUnit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      },
      "patch": {
        "operationId": "BuildingController_updateUnit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/units/{unitId}/invite-owner": {
      "post": {
        "operationId": "BuildingController_inviteOwner_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteOwnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/membership-requests": {
      "post": {
        "operationId": "BuildingController_requestMembership_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      },
      "get": {
        "operationId": "BuildingController_listMembershipRequests_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/membership-requests/{requestId}": {
      "patch": {
        "operationId": "BuildingController_resolveMembershipRequest_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveMembershipRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/manager": {
      "get": {
        "operationId": "BuildingController_getCurrentManager_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      },
      "patch": {
        "operationId": "BuildingController_changeManager_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeManagerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/manager/history": {
      "get": {
        "operationId": "BuildingController_getManagementHistory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/manager/end": {
      "post": {
        "operationId": "BuildingController_endManagement_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/units/{unitId}/ownership/history": {
      "get": {
        "operationId": "BuildingController_getOwnershipHistory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
    "/api/v1/buildings/{id}/units/{unitId}/ownership/transfer": {
      "post": {
        "operationId": "BuildingController_transferOwnership_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
...
    "/api/v1/buildings/{id}/units/{unitId}/tenancy/history": {
      "get": {
        "operationId": "BuildingController_getTenancyHistory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "building"
        ]
      }
    },
...
  },
  "info": {
    "title": "VielHome API",
    "description": "REST API for VielHome — see AIHandoff V2 for architecture and business rules.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SaveDraftDto": {
        "type": "object",
        "properties": {
          "step": {
            "type": "string",
            "enum": [
              "role_selection",
              "building_info",
              "address",
              "review"
            ]
          },
          "payload": {
            "type": "object"
          },
          "device": {
            "type": "string"
          }
        },
        "required": [
          "step",
          "payload"
        ]
      },
      "CreateUnitDto": {
        "type": "object",
        "properties": {
          "unitNumber": {
            "type": "string"
          },
          "blockId": {
            "type": "string"
          },
          "floorId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "RESIDENTIAL",
              "COMMERCIAL",
              "PARKING",
              "STORAGE"
            ]
          },
          "areaSqm": {
            "type": "number"
          }
        },
        "required": [
          "unitNumber"
        ]
      },
...
      "InviteOwnerDto": {
        "type": "object",
        "properties": {
          "ownerFullName": {
            "type": "string"
          },
          "ownerPhone": {
            "type": "string",
            "example": "+989121234567"
          }
        },
        "required": [
          "ownerFullName",
          "ownerPhone"
        ]
      },
...
      "ResolveMembershipRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}